Handle delete requests made via `POST` (#14)
If the user instructs FineUploader to use a `POST` request instead of a `DELETE` request, the handler will never clean up the file. E.g.,

```
deleteFile: {
        enabled: true,
        method: "POST"
    }
```

Automagically use the correct method to determine the file's `uuid` based on the server's `REQUEST_METHOD`.

Also, actually allow POST to be used to delete files.